EMP-762: use TlTaskLastUpdated as submitted date

+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/crm4/Crm4SubmissionModel.java
@@ -11,7 +11,7 @@ import java.util.Date;
 
 public class Crm4SubmissionModel {
     @JsonProperty("Usn")
-    public int usn;
+    public Long usn;
     @JsonProperty("Clientname")
     public String clientname;
     @JsonProperty("State")
